feat(ios): Create separate screens for examples
Keccak (Setup) and Keccak (Zkey) now both available

Also added mock RSA screen
